next try

+++ resources/views/bauprojekte.blade.php
@@ -621,10 +621,8 @@
     }
      
     function numericSort(a, b, rowA, rowB) {  
-      //a = (int)a;
-      //b = (int)b;
-      a = +a.substring(1); // remove $
-      b = +b.substring(1);
+      a = (int)+a.substring(1); // remove $
+      b = (int)+b.substring(1);
       if (a > b) return 1;
       if (a < b) return -1;
       return 0;
